requirements.txt
Django==5.2.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
gunicorn==23.0.0
Pillow==11.3.0
google-genai
